Implemented login tests and improved auth tests
