+++ src/factories/deleteFactory.js
@@ -1,0 +1,9 @@
+import connection from '../database/database';
+
+
+const cleanTableDatabase = async (tableName) => {
+  await connection.query(`DELETE FROM ${tableName};`);
+};
+
+
+export { cleanTableDatabase };
