+++ src/factories/fakerCpfFactory.js
@@ -1,0 +1,10 @@
+import faker from 'faker';
+
+
+const fakerCpf = () => String(faker.datatype.number({
+  min: 10000000000,
+  max: 99999999999
+}));
+
+
+export { fakerCpf };
